Как и предполагалось - выход за границы инта

+++ Practice1/Program.cs
@@ -6,17 +6,17 @@ using System.IO;
         static void Main(string[] args)
         {
             
-            int N;
+            Int64 N;
 
             StreamReader input = new StreamReader("input.txt");
             StreamWriter output = new StreamWriter("output.txt");
 
-            N = int.Parse(input.ReadLine());
-            int[] a = new int[N];
-            int min = -1;
-            int sum = 0;
+            N = Int64.Parse(input.ReadLine());
+            Int64[] a = new Int64[N];
+            Int64 min = -1;
+            Int64 sum = 0;
             for (int i = 0; i < N; i++)
-                a[i] = int.Parse(input.ReadLine());
+                a[i] = Int64.Parse(input.ReadLine());
 
             Array.Sort(a);
